Refactor some similar strategies
